+++ src/Region/RegionHandler.cc
@@ -4,7 +4,7 @@
    SPDX-License-Identifier: GPL-3.0-or-later
 */
 
-// RegionDataHandler.cc: handle requirements and data streams for regions
+// RegionHandler.cc: handle requirements and data streams for regions
 
 #include "RegionHandler.h"
 
@@ -283,8 +283,7 @@ void RegionHandler::RemoveFrame(int file_id) {
 
 bool RegionHandler::SetHistogramRequirements(
     int region_id, int file_id, std::shared_ptr<Frame> frame, const std::vector<CARTA::HistogramConfig>& configs) {
-    // Set histogram requirements for closed region
-
+    // Set histogram configurations for closed region
     if (configs.empty() && !RegionSet(region_id)) {
         // Frontend clears requirements after region removed, prevent error in log by returning true.
         return true;
@@ -303,16 +302,12 @@ bool RegionHandler::SetHistogramRequirements(
     // Save frame pointer
     _frames[file_id] = frame;
 
-    // Make HistogramConfig vector of requirements
-    std::vector<HistogramConfig> input_configs;
-    for (const auto& config : configs) {
-        HistogramConfig hist_config(config);
-        input_configs.push_back(hist_config);
+    if (_region_histograms.find(region_id) == _region_histograms.end()) {
+        _region_histograms[region_id] = std::make_unique<RegionHistogram>(region_id, file_id, configs);
+    } else {
+        _region_histograms[region_id]->SetConfigurations(file_id, configs);
     }
 
-    // Set requirements
-    ConfigId config_id(file_id, region_id);
-    _histogram_req[config_id].configs = input_configs;
     return true;
 }
 
@@ -518,10 +513,10 @@ void RegionHandler::UpdateNewSpectralRequirements(int region_id) {
 }
 
 bool RegionHandler::SetStatsRequirements(
-    int region_id, int file_id, std::shared_ptr<Frame> frame, const std::vector<CARTA::SetStatsRequirements_StatsConfig>& stats_configs) {
+    int region_id, int file_id, std::shared_ptr<Frame> frame, const std::vector<CARTA::SetStatsRequirements_StatsConfig>& configs) {
     // Set stats data requirements for closed region
 
-    if (stats_configs.empty() && !RegionSet(region_id)) {
+    if (configs.empty() && !RegionSet(region_id)) {
         // frontend clears requirements after region removed, prevent error in log
         return true;
     }
@@ -539,17 +534,20 @@ bool RegionHandler::SetStatsRequirements(
     // Save frame pointer
     _frames[file_id] = frame;
 
-    // Set requirements
-    ConfigId config_id(file_id, region_id);
-    _stats_req[config_id].stats_configs = stats_configs;
+    // Set configurations
+    if (_region_statistics.find(region_id) == _region_statistics.end()) {
+        _region_statistics[region_id] = std::make_unique<RegionStatistics>(region_id, file_id, configs);
+    } else {
+        _region_statistics[region_id]->SetConfigurations(file_id, configs);
+    }
     return true;
 }
 
 void RegionHandler::RemoveRegionRequirementsCache(int region_id) {
     // Clear requirements and cache for a specific region or for all regions when closed
     if (region_id == ALL_REGIONS) {
-        _histogram_req.clear();
-        _stats_req.clear();
+        _region_histograms.clear();
+        _region_statistics.clear();
 
         std::unique_lock<std::mutex> spectral_lock(_spectral_mutex);
         _spectral_req.clear();
@@ -559,24 +557,24 @@ void RegionHandler::RemoveRegionRequirementsCache(int region_id) {
         _spatial_req.clear();
         spatial_lock.unlock();
 
-        _histogram_cache.clear();
         _spectral_cache.clear();
-        _stats_cache.clear();
 
         std::unique_lock pv_cut_lock(_pv_cut_mutex);
         _pv_preview_cuts.clear();
         std::unique_lock pv_cube_lock(_pv_cube_mutex);
         _pv_preview_cubes.clear();
     } else {
-        // Iterate through requirements and remove those for given region_id
-        for (auto it = _histogram_req.begin(); it != _histogram_req.end();) {
-            if ((*it).first.region_id == region_id) {
-                it = _histogram_req.erase(it);
-            } else {
-                ++it;
-            }
+        // Remove histograms for given region_id
+        if (_region_histograms.find(region_id) != _region_histograms.end()) {
+            _region_histograms.erase(region_id);
         }
 
+        // Remove statistics for given region_id
+        if (_region_statistics.find(region_id) != _region_statistics.end()) {
+            _region_statistics.erase(region_id);
+        }
+
+        // Iterate through requirements and remove those for given region_id
         std::unique_lock<std::mutex> spectral_lock(_spectral_mutex);
         for (auto it = _spectral_req.begin(); it != _spectral_req.end();) {
             if ((*it).first.region_id == region_id) {
@@ -586,14 +584,6 @@ void RegionHandler::RemoveRegionRequirementsCache(int region_id) {
             }
         }
         spectral_lock.unlock();
-
-        for (auto it = _stats_req.begin(); it != _stats_req.end();) {
-            if ((*it).first.region_id == region_id) {
-                it = _stats_req.erase(it);
-            } else {
-                ++it;
-            }
-        }
 
         std::unique_lock<std::mutex> spatial_lock(_spatial_mutex);
         for (auto it = _spatial_req.begin(); it != _spatial_req.end();) {
@@ -605,25 +595,9 @@ void RegionHandler::RemoveRegionRequirementsCache(int region_id) {
         }
         spatial_lock.unlock();
 
-        for (auto it = _histogram_cache.begin(); it != _histogram_cache.end();) {
-            if ((*it).first.region_id == region_id) {
-                it = _histogram_cache.erase(it);
-            } else {
-                ++it;
-            }
-        }
-
         for (auto it = _spectral_cache.begin(); it != _spectral_cache.end();) {
             if ((*it).first.region_id == region_id) {
                 it = _spectral_cache.erase(it);
-            } else {
-                ++it;
-            }
-        }
-
-        for (auto it = _stats_cache.begin(); it != _stats_cache.end();) {
-            if ((*it).first.region_id == region_id) {
-                it = _stats_cache.erase(it);
             } else {
                 ++it;
             }
@@ -647,8 +621,9 @@ void RegionHandler::RemoveRegionRequirementsCache(int region_id) {
 void RegionHandler::RemoveFileRequirementsCache(int file_id) {
     // Clear requirements and cache for a specific file or for all files when closed
     if (file_id == ALL_FILES) {
-        _histogram_req.clear();
-        _stats_req.clear();
+        _region_histograms.clear();
+        _region_statistics.clear();
+
         std::unique_lock<std::mutex> spectral_lock(_spectral_mutex);
         _spectral_req.clear();
         spectral_lock.unlock();
@@ -656,24 +631,24 @@ void RegionHandler::RemoveFileRequirementsCache(int file_id) {
         _spatial_req.clear();
         spatial_lock.unlock();
 
-        _histogram_cache.clear();
         _spectral_cache.clear();
-        _stats_cache.clear();
 
         std::unique_lock pv_cut_lock(_pv_cut_mutex);
         _pv_preview_cuts.clear();
         std::unique_lock pv_cube_lock(_pv_cube_mutex);
         _pv_preview_cubes.clear();
     } else {
-        // Iterate through requirements and remove those for given file_id
-        for (auto it = _histogram_req.begin(); it != _histogram_req.end();) {
-            if ((*it).first.file_id == file_id) {
-                it = _histogram_req.erase(it);
-            } else {
-                ++it;
-            }
+        // Remove histograms for given file_id
+        for (const auto& [_, region_histogram] : _region_histograms) {
+            region_histogram->ClearFileConfigsCache(file_id);
         }
 
+        // Remove statistics for given file_id
+        for (const auto& [_, region_statistics] : _region_statistics) {
+            region_statistics->ClearFileConfigsCache(file_id);
+        }
+
+        // Iterate through requirements and remove those for given file_id
         std::unique_lock<std::mutex> spectral_lock(_spectral_mutex);
         for (auto it = _spectral_req.begin(); it != _spectral_req.end();) {
             if ((*it).first.file_id == file_id) {
@@ -683,14 +658,6 @@ void RegionHandler::RemoveFileRequirementsCache(int file_id) {
             }
         }
         spectral_lock.unlock();
-
-        for (auto it = _stats_req.begin(); it != _stats_req.end();) {
-            if ((*it).first.file_id == file_id) {
-                it = _stats_req.erase(it);
-            } else {
-                ++it;
-            }
-        }
 
         std::unique_lock<std::mutex> spatial_lock(_spatial_mutex);
         for (auto it = _spatial_req.begin(); it != _spatial_req.end();) {
@@ -703,25 +670,9 @@ void RegionHandler::RemoveFileRequirementsCache(int file_id) {
         spatial_lock.unlock();
 
         // Iterate through cache and remove those for given file_id
-        for (auto it = _histogram_cache.begin(); it != _histogram_cache.end();) {
-            if ((*it).first.file_id == file_id) {
-                it = _histogram_cache.erase(it);
-            } else {
-                ++it;
-            }
-        }
-
         for (auto it = _spectral_cache.begin(); it != _spectral_cache.end();) {
             if ((*it).first.file_id == file_id) {
                 it = _spectral_cache.erase(it);
-            } else {
-                ++it;
-            }
-        }
-
-        for (auto it = _stats_cache.begin(); it != _stats_cache.end();) {
-            if ((*it).first.file_id == file_id) {
-                it = _stats_cache.erase(it);
             } else {
                 ++it;
             }
@@ -751,20 +702,17 @@ void RegionHandler::RemoveFileRequirementsCache(int file_id) {
 
 void RegionHandler::ClearRegionCache(int region_id) {
     // Remove cached data when region changes
-    for (auto& hcache : _histogram_cache) {
-        if (hcache.first.region_id == region_id) {
-            hcache.second.ClearHistograms();
-        }
+    if (_region_histograms.find(region_id) != _region_histograms.end()) {
+        _region_histograms[region_id]->ClearCache();
     }
     for (auto& spcache : _spectral_cache) {
         if (spcache.first.region_id == region_id) {
             spcache.second.ClearProfiles();
         }
     }
-    for (auto& stcache : _stats_cache) {
-        if (stcache.first.region_id == region_id) {
-            stcache.second.ClearStats();
-        }
+
+    if (_region_statistics.find(region_id) != _region_statistics.end()) {
+        _region_statistics[region_id]->ClearCache();
     }
 }
 
@@ -1186,7 +1134,7 @@ bool RegionHandler::CalculatePvPreviewImage(int frame_id, int preview_id, bool q
         auto box_mask = _regions.at(box_region_id)->GetImageRegionMask(frame_id);
         RemoveRegion(box_region_id);
 
-        // Use PvPreviewCube to calculate profile with lcregion and mask
+        // Use PvPreviewCube to calculate profile with lc_region and mask
         std::vector<float> profile;
         double max_num_pixels(0.0);
         std::string message;
@@ -1508,177 +1456,62 @@ bool RegionHandler::FitImage(const CARTA::FittingRequest& fitting_request, CARTA
 // region_id < 0 file_id < 0    not allowed (all regions for all files?)
 // region_id = 0                not allowed (cursor region handled by Frame)
 
-// ***** Fill histogram *****
-
-bool RegionHandler::FillRegionHistogramData(
-    std::function<void(CARTA::RegionHistogramData histogram_data)> region_histogram_callback, int region_id, int file_id) {
-    // Fill histogram data for given region and file
+bool RegionHandler::FillRegionHistogramData(std::function<void(CARTA::RegionHistogramData histogram_data)> cb, int region_id, int file_id) {
     if (!RegionFileIdsValid(region_id, file_id, true)) {
         return false;
     }
 
-    bool message_filled(false);
-    if (region_id > 0) {
-        // Fill histograms for specific region with file_id requirement (specific file_id or all files)
-        std::unordered_map<ConfigId, RegionHistogramConfig, ConfigIdHash> region_configs = _histogram_req;
-        for (auto& region_config : region_configs) {
-            if ((region_config.first.region_id == region_id) && ((region_config.first.file_id == file_id) || (file_id == ALL_FILES))) {
-                if (region_config.second.configs.empty()) { // no requirements
-                    continue;
-                }
-                int config_file_id = region_config.first.file_id;
-                if (!RegionFileIdsValid(region_id, config_file_id)) { // check specific ids
-                    continue;
-                }
+    if ((region_id > 0) && (_region_histograms.find(region_id) == _region_histograms.end())) {
+        return false;
+    }
 
-                // return histograms for this requirement
-                std::vector<CARTA::RegionHistogramData> histogram_messages;
-                std::vector<HistogramConfig> histogram_configs = region_config.second.configs;
-                if (GetRegionHistogramData(region_id, config_file_id, histogram_configs, histogram_messages)) {
-                    for (const auto& histogram_message : histogram_messages) {
-                        region_histogram_callback(histogram_message); // send histogram data with respect to stokes
-                    }
-                    message_filled = true;
-                }
-            }
+    bool success(false);
+    for (const auto& [hist_region_id, region_histogram] : _region_histograms) {
+        // Find histogram configurations with region_id and file_id
+        if ((region_id > 0) && (hist_region_id != region_id)) {
+            continue;
         }
-    } else {
-        // (region_id < 0) Fill histograms for all regions with specific file_id requirement
-        std::unordered_map<ConfigId, RegionHistogramConfig, ConfigIdHash> region_configs = _histogram_req;
-        for (auto& region_config : region_configs) {
-            if (region_config.first.file_id == file_id) {
-                if (region_config.second.configs.empty()) { // requirements removed
-                    continue;
-                }
-                int config_region_id(region_config.first.region_id);
-                if (!RegionFileIdsValid(config_region_id, file_id)) { // check specific ids
+
+        auto config_file_ids = region_histogram->GetConfigFileIds(file_id);
+        if (config_file_ids.empty()) {
+            continue;
+        }
+
+        for (int hist_file_id : config_file_ids) {
+            // Get histogram for specific region id and file id
+            if (!RegionFileIdsValid(hist_region_id, hist_file_id)) {
+                continue;
+            }
+
+            auto frame = _frames.at(hist_file_id);
+
+            std::vector<HistogramConfig> histogram_configs;
+            if (!region_histogram->GetConfigurations(hist_file_id, histogram_configs)) {
+                continue;
+            }
+
+            for (auto& histogram_config : histogram_configs) {
+                // Create data message for each configuration
+                int stokes(0);
+                if (!frame->GetStokesTypeIndex(histogram_config.coordinate, stokes)) {
                     continue;
                 }
 
-                // return histograms for this requirement
-                std::vector<HistogramConfig> histogram_configs = region_config.second.configs;
-                std::vector<CARTA::RegionHistogramData> histogram_messages;
-                if (GetRegionHistogramData(config_region_id, file_id, histogram_configs, histogram_messages)) {
-                    for (const auto& histogram_message : histogram_messages) {
-                        region_histogram_callback(histogram_message); // send histogram data with respect to stokes
-                    }
-                    message_filled = true;
+                int z = (histogram_config.channel == CURRENT_Z ? frame->CurrentZ() : histogram_config.channel);
+                AxisRange z_range(z);
+                StokesSource stokes_source(stokes, z_range);
+                std::shared_ptr<casacore::LCRegion> lcregion = ApplyRegionToFile(hist_region_id, hist_file_id, stokes_source);
+                CARTA::RegionHistogramData histogram_data_message;
+
+                if (region_histogram->GetRegionHistogramData(
+                        hist_file_id, frame, histogram_config, lcregion, stokes_source, histogram_data_message)) {
+                    cb(histogram_data_message);
+                    success = true;
                 }
             }
         }
     }
-    return message_filled;
-}
-
-bool RegionHandler::GetRegionHistogramData(
-    int region_id, int file_id, std::vector<HistogramConfig>& configs, std::vector<CARTA::RegionHistogramData>& histogram_messages) {
-    // Fill stats message for given region, file
-    Timer t;
-
-    // Set channel range is the current channel
-    int z(_frames.at(file_id)->CurrentZ());
-    AxisRange z_range(z);
-
-    // Stokes type to be determined in each of histogram configs
-    int stokes;
-
-    // Flags for calculations
-    bool have_basic_stats(false);
-
-    // Reuse the image region for each histogram
-    StokesRegion stokes_region;
-    std::shared_ptr<casacore::LCRegion> lc_region;
-
-    // Reuse data with respect to stokes and stats for each histogram; results depend on num_bins
-    std::unordered_map<int, std::vector<float>> data;
-    BasicStats<float> stats;
-
-    for (auto& hist_config : configs) {
-        // check for cancel
-        if (!RegionFileIdsValid(region_id, file_id)) {
-            return false;
-        }
-
-        // Get stokes index
-        if (!_frames.at(file_id)->GetStokesTypeIndex(hist_config.coordinate, stokes)) {
-            continue;
-        }
-
-        // Set histogram fields
-        auto histogram_message = Message::RegionHistogramData(file_id, region_id, z, stokes, 1.0, hist_config);
-
-        // Get image region
-        if (!ApplyRegionToFile(region_id, file_id, z_range, stokes, lc_region, stokes_region)) {
-            // region outside image, send default histogram
-            auto* default_histogram = histogram_message.mutable_histograms();
-            std::vector<int> histogram_bins(1, 0);
-            FillHistogram(default_histogram, 1, 0.0, 0.0, histogram_bins, DOUBLE_NAN, DOUBLE_NAN);
-            continue;
-        }
-
-        // number of bins may be set or calculated
-        int num_bins(hist_config.num_bins);
-        if (num_bins == AUTO_BIN_SIZE) {
-            casacore::IPosition region_shape = _frames.at(file_id)->GetRegionShape(stokes_region);
-            num_bins = int(std::max(sqrt(region_shape(0) * region_shape(1)), 2.0));
-        }
-
-        // Key for cache
-        CacheId cache_id = CacheId(file_id, region_id, stokes, z);
-
-        // check cache
-        if (_histogram_cache.count(cache_id)) {
-            have_basic_stats = _histogram_cache[cache_id].GetBasicStats(stats);
-            if (have_basic_stats) {
-                // Set histogram bounds
-                auto bounds = hist_config.GetBounds(stats);
-                Histogram hist;
-                if (_histogram_cache[cache_id].GetHistogram(num_bins, bounds, hist)) {
-                    auto* histogram = histogram_message.mutable_histograms();
-                    FillHistogram(histogram, stats, hist);
-
-                    // Fill in the cached message
-                    histogram_messages.emplace_back(histogram_message);
-                    continue;
-                }
-            }
-        }
-
-        // Calculate stats and/or histograms, not in cache
-        // Get data in region
-        if (!data.count(stokes)) {
-            if (!_frames.at(file_id)->GetRegionData(stokes_region, data[stokes])) {
-                spdlog::error("Failed to get data in the region!");
-                return false;
-            }
-        }
-
-        // Calculate and cache stats
-        if (!have_basic_stats) {
-            CalcBasicStats(stats, data[stokes].data(), data[stokes].size());
-            _histogram_cache[cache_id].SetBasicStats(stats);
-            have_basic_stats = true;
-        }
-
-        // Set histogram bounds
-        Bounds bounds = hist_config.GetBounds(stats);
-
-        // Calculate and cache histogram for number of bins
-        Histogram histo = CalcHistogram(num_bins, bounds, data[stokes].data(), data[stokes].size());
-        _histogram_cache[cache_id].SetHistogram(num_bins, histo);
-
-        // Complete Histogram submessage
-        auto* histogram = histogram_message.mutable_histograms();
-        FillHistogram(histogram, stats, histo);
-
-        // Fill in the final result
-        histogram_messages.emplace_back(histogram_message);
-    }
-
-    auto dt = t.Elapsed();
-    spdlog::performance("Fill region histogram in {:.3f} ms at {:.3f} MPix/s", dt.ms(), (float)stats.num_pixels / dt.us());
-
-    return true;
+    return success;
 }
 
 // ***** Fill spectral profile *****
@@ -2042,142 +1875,66 @@ bool RegionHandler::GetRegionSpectralData(int region_id, int file_id, const Axis
     return true;
 }
 
-// ***** Fill stats data *****
-
 bool RegionHandler::FillRegionStatsData(std::function<void(CARTA::RegionStatsData stats_data)> cb, int region_id, int file_id) {
     // Fill stats data for given region and file
     if (!RegionFileIdsValid(region_id, file_id, true)) {
         return false;
     }
 
-    bool message_filled(false);
+    if ((region_id > 0) && (_region_statistics.find(region_id) == _region_statistics.end())) {
+        return false;
+    }
 
-    auto send_stats_results = [&](int region_id, int file_id, std::vector<CARTA::SetStatsRequirements_StatsConfig> stats_configs) {
-        for (auto stats_config : stats_configs) {
-            // Get stokes index
-            int stokes;
-            if (!_frames.at(file_id)->GetStokesTypeIndex(stats_config.coordinate(), stokes)) {
+    bool success(false);
+
+    for (const auto& [stats_region_id, region_statistics] : _region_statistics) {
+        if ((region_id > 0) && (stats_region_id != region_id)) {
+            continue;
+        }
+
+        auto config_file_ids = region_statistics->GetConfigFileIds(file_id);
+        if (config_file_ids.empty()) {
+            continue;
+        }
+
+        for (int stats_file_id : config_file_ids) {
+            // Get statistics for specific region id and file id
+            if (!RegionFileIdsValid(stats_region_id, stats_file_id)) {
                 continue;
             }
 
-            // Set required stats types
-            std::vector<CARTA::StatsType> required_stats;
-            for (int i = 0; i < stats_config.stats_types_size(); ++i) {
-                required_stats.push_back(stats_config.stats_types(i));
+            auto frame = _frames.at(stats_file_id);
+
+            std::vector<CARTA::SetStatsRequirements_StatsConfig> stats_configs;
+            if (!region_statistics->GetConfigurations(stats_file_id, stats_configs)) {
+                continue;
             }
 
-            // Send stats results
-            CARTA::RegionStatsData stats_message;
-            if (GetRegionStatsData(region_id, file_id, stokes, required_stats, stats_message)) {
-                cb(stats_message); // send stats data with respect to stokes
-                message_filled = true;
-            }
-        }
-    };
-
-    if (region_id > 0) {
-        // Fill stats data for specific region with file_id requirement (specific file_id or all files)
-        std::unordered_map<ConfigId, RegionStatsConfig, ConfigIdHash> region_configs = _stats_req;
-        for (auto& region_config : region_configs) {
-            if ((region_config.first.region_id == region_id) && ((region_config.first.file_id == file_id) || (file_id == ALL_FILES))) {
-                if (region_config.second.stats_configs.empty()) { // no requirements
-                    continue;
-                }
-                int config_file_id(region_config.first.file_id);
-                if (!RegionFileIdsValid(region_id, config_file_id)) { // check specific ids
-                    continue;
+            // Create data message for each configuration
+            for (auto& stats_config : stats_configs) {
+                // Get stokes and channel from config
+                int stokes(0);
+                if (!frame->GetStokesTypeIndex(stats_config.coordinate(), stokes)) {
+                    continue; // invalid image/computed Stokes
                 }
 
-                // return stats for this requirement
-                send_stats_results(region_id, config_file_id, region_config.second.stats_configs);
-            }
-        }
-    } else {
-        // (region_id < 0) Fill stats data for all regions with specific file_id requirement
-        // Find requirements with file_id
-        std::unordered_map<ConfigId, RegionStatsConfig, ConfigIdHash> region_configs = _stats_req;
-        for (auto& region_config : region_configs) {
-            if (region_config.first.file_id == file_id) {
-                if (region_config.second.stats_configs.empty()) { // no requirements
-                    continue;
-                }
-                int config_region_id(region_config.first.region_id);
-                if (!RegionFileIdsValid(config_region_id, file_id)) { // check specific ids
-                    continue;
-                }
+                // Get LCRegion
+                int z(_frames.at(stats_file_id)->CurrentZ());
+                AxisRange z_range(z);
+                std::shared_ptr<casacore::LCRegion> lc_region;
+                StokesRegion stokes_region;
+                ApplyRegionToFile(stats_region_id, stats_file_id, z_range, stokes, lc_region, stokes_region);
+                CARTA::RegionStatsData stats_data_message;
 
-                // return stats for this requirement
-                send_stats_results(config_region_id, file_id, region_config.second.stats_configs);
+                if (region_statistics->GetRegionStatsData(stats_file_id, frame, stats_config, stokes_region, stats_data_message)) {
+                    cb(stats_data_message);
+                    success = true;
+                }
             }
         }
     }
-    return message_filled;
-}
 
-bool RegionHandler::GetRegionStatsData(
-    int region_id, int file_id, int stokes, const std::vector<CARTA::StatsType>& required_stats, CARTA::RegionStatsData& stats_message) {
-    // Fill stats message for given region, file
-    Timer t;
-
-    int z(_frames.at(file_id)->CurrentZ());
-    stokes = (stokes == CURRENT_STOKES) ? _frames.at(file_id)->CurrentStokes() : stokes;
-
-    // Start filling message
-    stats_message.set_file_id(file_id);
-    stats_message.set_region_id(region_id);
-    stats_message.set_channel(z);
-    stats_message.set_stokes(stokes);
-
-    // Check cache
-    CacheId cache_id = CacheId(file_id, region_id, stokes, z);
-    if (_stats_cache.count(cache_id)) {
-        std::map<CARTA::StatsType, double> stats_results;
-        if (_stats_cache[cache_id].GetStats(stats_results)) {
-            FillStatistics(stats_message, required_stats, stats_results);
-            return true;
-        }
-    }
-
-    // Get region
-    AxisRange z_range(z);
-    StokesRegion stokes_region;
-    std::shared_ptr<casacore::LCRegion> lc_region;
-    if (!ApplyRegionToFile(region_id, file_id, z_range, stokes, lc_region, stokes_region)) {
-        // region outside image: NaN results
-        std::map<CARTA::StatsType, double> stats_results;
-        for (const auto& carta_stat : required_stats) {
-            if (carta_stat == CARTA::StatsType::NumPixels) {
-                stats_results[carta_stat] = 0.0;
-            } else {
-                stats_results[carta_stat] = DOUBLE_NAN;
-            }
-        }
-        FillStatistics(stats_message, required_stats, stats_results);
-        // cache results
-        _stats_cache[cache_id] = StatsCache(stats_results);
-        return true;
-    }
-
-    // calculate stats
-    bool per_z(false);
-    std::map<CARTA::StatsType, std::vector<double>> stats_map;
-    if (_frames.at(file_id)->GetRegionStats(stokes_region, required_stats, per_z, stats_map)) {
-        // convert vector to single value in map
-        std::map<CARTA::StatsType, double> stats_results;
-        for (auto& value : stats_map) {
-            stats_results[value.first] = value.second[0];
-        }
-
-        // add values to message
-        FillStatistics(stats_message, required_stats, stats_results);
-        // cache results
-        _stats_cache[cache_id] = StatsCache(stats_results);
-
-        spdlog::performance("Fill region stats in {:.3f} ms", t.Elapsed().ms());
-        return true;
-    }
-
-    return false;
+    return success;
 }
 
 bool RegionHandler::FillPointSpatialProfileData(int file_id, int region_id, std::vector<CARTA::SpatialProfileData>& spatial_data_vec) {
@@ -2196,13 +1953,13 @@ bool RegionHandler::FillPointSpatialProfileData(int file_id, int region_id, std:
     }
 
     // Map a region (region_id) to an image (file_id)
-    auto lcregion = ApplyRegionToFile(region_id, file_id);
-    if (!lcregion) {
+    auto lc_region = ApplyRegionToFile(region_id, file_id);
+    if (!lc_region) {
         return false;
     }
 
     // Get point region
-    casacore::IPosition origin = lcregion->boundingBox().start();
+    casacore::IPosition origin = lc_region->boundingBox().start();
     PointXy point(origin(0), origin(1));
 
     // Get profile for point region